third_party/WebKit/LayoutTests/sensor/idl-AbsoluteOrientationSensor.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script>

test(function() {
  // Test that AbsoluteOrientationSensor interface exists
  assert_true('AbsoluteOrientationSensor' in window);
}, 'Test that the AbsoluteOrientationSensor interface is present.');

</script>
